+++ lib/main.ts
@@ -1,0 +1,3 @@
+import { Collapse } from './Collapse.tsx'
+
+export { Collapse }