+++ node-tools/clean-dist.js
@@ -1,0 +1,7 @@
+import fs from 'fs'
+
+function removeFolderContents(dir) {
+  fs.rmSync(dir, { recursive: true, force: true })
+}
+
+removeFolderContents('dist')