+++ node-tools/copy-files.js
@@ -1,0 +1,4 @@
+import { copyFile } from 'copy-file'
+
+await copyFile('types/index.d.ts', 'dist/react-collapse.d.ts')
+console.log('** File copied')