+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # react-collapse changelog
 
+## 3.0.0
+
+_2024_
+
+- Change API. Use this instead
+  - `import { Collapse } from '@kunukn/react-collapse'`
+- Support SSR hydration.
+- React 16.8+ version
+- Migrate library to use Vite.
+
 ## 2.2.10
 
 Support React 18